Disable incremental invalidation for SVG
Incremental invalidation is a feature where size changes (e.g.,
shrinking an object's height) can only invalidate the shrunken region
instead of the entire object. This can lead to underinvalidation bugs
if the object is not fully opaque, such as the case of rounded corners.

This patch switches all incremental invalidations to be full
invalidations.

BUG=619108

Review-Url: https://codereview.chromium.org/2062533002
Cr-Commit-Position: refs/heads/master@{#399384}