Add "Checking" spinner for server-side shipping address validation.
PaymentRequest should  let the user know that their section is being
verified. This is used, for example, when the website is verifying
availability of shipping options to the shipping address that the user
has selected. If an address is not supported, an error message should be
shown. The unsupported address should be disabled.

BUG=619318

Review-Url: https://codereview.chromium.org/2059073002
Cr-Commit-Position: refs/heads/master@{#399612}