LayoutObject::paintInvalidationParent() to fix multicol spanner invalidation
This was LayoutObject::parentCrossingFrameBoundaries(). Renamed and
added special rule for column spanner, and used for both finding
painting layer and marking ancestor chain for paint invalidation.

Previously we marked ancestor chain for paint invalidation through
the normal tree order, causing the placeholder not marked for
paint invalidation and missing walking to the spanner. If dcheck
was enabled, it would crash because the paint invalidation flags
of the spanner is not cleared.

BUG=619263
TEST=fast/repaint/multicol-spanner-background-color.html

Review-Url: https://codereview.chromium.org/2063543002
Cr-Commit-Position: refs/heads/master@{#399614}