DatabaseTracker: Use 'raw' string as key for origin
Prior to r382660 the WebSQL implementation used DatabaseIdentifiers, and
a serialization was used as the key in maps. This was changed to take
SecurityOrigins instead and the serialization via toString(). This
introduced additional special cases e.g. "null" for file:// origins in
some cases, leading to mismatches between setting and lookup in the
map depending on when the SecurityOrigin was minted. To fix, use
toRawString() within the DatabaseTracker which matches the old behavior.

Similarly, use toRawString() when serializing an origin for posting
across threads.

BUG=618381
R=michaeln@chromium.org

Review-Url: https://codereview.chromium.org/2050803005
Cr-Commit-Position: refs/heads/master@{#399743}