Switch SizeF from using fmaxf() to direct comparisons.
This is a prerequisite for changing SizeF's constructors to be constexpr.

The comparison must be "width >= 0 ? width : 0" and not "width < 0 ? 0 : width"
in order to maintain consistency with how fmaxf() handled NaN, which some tests
rely on.  (We want to clamp NaN to 0 instead of allowing it through.)

For consistency, changed BoxF to use the same comparison form, so it will handle
NaN the same way.

Tested the thread_times.key_mobile_sites_smooth perf test on Android with this
patch and verified that on most bots (except Nexus 6 and FYI bots) this was
neutral-to-positive.  See https://codereview.chromium.org/2066893002/#ps20001 .

BUG=none
TEST=none

Review-Url: https://codereview.chromium.org/2060313004
Cr-Commit-Position: refs/heads/master@{#399799}