[sync] Search engine shortcuts get underscores at the end after sync
Currently, search engines with matching prepopulate_id are treated as distinct if their keywords don't match. This can cause multiple copies of prepopulated search engines to be created; one copy with the original keyword and another copy with the modified keyword.

In the linked bug, the problematic behavior has created a different default search engine on each client. Deleting the non-default on client 1 is an attempt to delete client 2's default. There is sync code that assumes this was a mistake and resurrects the local default by appending an underscore.

This CL treats matching prepopulate_id as a conflict, and resolves it in favor of the remote search engine if the local engine isn't yet known to sync and was modified less recently. Note that this won't fix existing clients, whose duplicated engines will be known to sync already.

R=pkasting@chromium.org, zea@chromium.org

BUG=613108

Review-Url: https://codereview.chromium.org/2067723002
Cr-Commit-Position: refs/heads/master@{#400017}