Adjust BackgroundWith1PxBorder corner radius to scale properly.
The ".5" in the existing radius was intended to get to the middle of the stroke,
but that portion of the radius value was in px and not DIP, so scaling the whole
thing wasn't correct.  Instead, set the corner radius to the value of the "inner
radius", then add .5 after scaling.

This fixes a problem where, at larger scale factors, the focus ring's curvature
was slightly different than the edge's curvature.  It also very slightly changes
the appearance of the corner (no one would notice without zooming in).

BUG=612758
TEST=none

Review-Url: https://codereview.chromium.org/2066303003
Cr-Commit-Position: refs/heads/master@{#400089}