Give MD text button text medium weight.
This will not always work because there may not be a
medium-weight version of the system font available. For
example, my stock Ubuntu config uses Sans as the default
font and there's no Sans Medium, so it remains Sans. If I
change my system font to the font "Ubuntu", which does have
a medium variant, I get a bolder font (note that Ubuntu
Medium is heavier than you might expect for "medium").

This works on CrOS as well because we package Roboto-
Medium. But to get a desktop build of CrOS to work on
Ubuntu, you must 'sudo apt-get install fonts-roboto'.
There's a bug[1] where Roboto Medium is missing from this
package on 14.04, so you have to manually install the newer
package (see link).

[1] https://bugs.launchpad.net/ubuntu/+source/fonts-android/+bug/1580341

BUG=571500

Review-Url: https://codereview.chromium.org/2065513005
Cr-Commit-Position: refs/heads/master@{#400178}