Fix glass extension on MD for Win 7 and below.
This also makes visible (but minor) changes on non-MD (the color of the
innermost pixel of client edge shifts slightly).  This is because we didn't
previously inset the glass enough for the light pixel of edge to be completely
inside the semitransparent client edge region.

This also fixes the glass extension if we were to scale up larger than 200%
somehow.

All of these problems were caused by incorrectly computing the necessary
extension.  The biggest problem was that the code did not account for the
innermost 2 px of glass being a very light and then very dark pixel.  We need to
inset by that amount additionally.

The code also used the wrong mechanism to convert DIPs to pixels, which I think
would have caused bugs on systems with multiple monitors with different DPI.

BUG=614146
TEST=Start Chrome on Win 7 and verify there are no bright or dark lines across the window just above the bottom of the tabstrip.

Review-Url: https://codereview.chromium.org/2056583002
Cr-Commit-Position: refs/heads/master@{#400316}