Add BlimpContents[,Observer] and BlimpNavigationController
This CL adds the public API for //blimp/client/core and adds a simple
implementation with a test for the core functionality.

The setup of BlimpContents and related classes is intentionally
similar to how the //content API is organized.

Users of this API are supposed to depend on the code in
//blimp/client/core/public, and not the implementation which lives
in //blimp/client/core directly.

BUG=611103

Review-Url: https://codereview.chromium.org/2040813004
Cr-Commit-Position: refs/heads/master@{#400456}