Don't set a default action for ToolbarActionsBarBubbles.
This isn't totally correct; some of these bubbles should have defaults
and some should not. But this was the behavior previous to dc0d234f7c87
so the more conservative approach is to revert to the old behavior and
later change/improve individual bubbles that should have defaults by
adding them.

This patch is designed to be easy/safe to merge to M52.

BUG=621122

Review-Url: https://codereview.chromium.org/2076993002
Cr-Commit-Position: refs/heads/master@{#400503}