Service worker: Remove unneeded early stop in FinishRequest
This is a vestige of older code added in https://crrev.com/0c171c6df9b7fbb
Previously Doom() called StopIfIdle() rather than Stop(). Now (as of
https://crrev.com/b34f1e5b97667e) Doom() terminates immediately, with the one
exception when DevTools is attached. In that case, we stop immediately once
DevTools is detached.

There are two callsites that set status to REDUNDANT directly. This patch
changes one to call Stop() first, as per spec. It's not necessary to call
Doom() because the new version has already been stored which puts the old
version's resources in the purged list.

This is minor cleanup in anticipation of some changes for
https://crbug.com/616331

BUG=616331,448670
TEST=third_party/WebKit/Tools/Scripts/run-webkit-tests --repeat-each=50 http/tests/serviceworker/register-same-scope-different-script-url.html

Review-Url: https://codereview.chromium.org/2077843002
Cr-Commit-Position: refs/heads/master@{#400636}