Update size of Google Logo on Android NTP.
The logo is now 64dp high on <600dp screens, and 96dp on bigger screens.

Because the default logo is never upscaled, using smaller assets while keeping the size of the view the same works, and ensures Doodles will not get downscaled.

Screenshots on tablet/phone: https://imgur.com/a/ntrQj

BUG=612156

Review-Url: https://codereview.chromium.org/2068883003
Cr-Commit-Position: refs/heads/master@{#400769}